fix refl

+++ Pulsar/Source/Pulsar/include/Pulsar/Assets/Texture.h
@@ -18,7 +18,7 @@ namespace pulsar
         virtual int32_t GetWidth() const = 0;
         virtual int32_t GetHeight() const = 0;
         virtual Vector2i GetSize2di() const { return { this->GetWidth(), this->GetHeight() }; }
-        virtual Vector2f GetSize2df() const { return Vector2f(this->GetWidth(), this->GetHeight()); }
+        virtual Vector2f GetSize2df() const { return Vector2f((float)this->GetWidth(), (float)this->GetHeight()); }
     protected:
         uint32_t tex_id_ = 0;
     };
